pyproject.toml
[project]
name = "windwhisper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "jupyter>=1.1.1",
    "keyboard>=0.13.5",
    "numpy>=2.0.2",
    "openai-whisper>=20240930",
    "playsound==1.2.2",
    "pynput>=1.7.7",
    "pyperclip>=1.9.0",
    "scipy>=1.13.1",
    "sounddevice>=0.5.1",
]
